+++ framework/PageTable.tsx
@@ -1,52 +1,20 @@
 import {
     Button,
-    ClipboardCopy,
     EmptyState,
     EmptyStateBody,
     EmptyStateIcon,
     EmptyStateSecondaryActions,
     EmptyStateVariant,
-    Flex,
-    FlexItem,
-    InputGroup,
-    Label,
-    LabelGroup,
-    OnPerPageSelect,
-    OnSetPage,
-    Pagination,
-    PaginationVariant,
-    SelectOption,
     Skeleton,
     Split,
     SplitItem,
-    TextInputGroup,
-    TextInputGroupMain,
-    TextInputGroupUtilities,
     Title,
-    Toolbar,
-    ToolbarContent,
-    ToolbarFilter,
-    ToolbarGroup,
-    ToolbarItem,
-    ToolbarToggleGroup,
-    Tooltip,
-    Truncate,
 } from '@patternfly/react-core'
-import {
-    ArrowRightIcon,
-    ColumnsIcon,
-    ExclamationCircleIcon,
-    FilterIcon,
-    PlusCircleIcon,
-    SearchIcon,
-    TimesIcon,
-} from '@patternfly/react-icons'
+import { ExclamationCircleIcon, PlusCircleIcon, SearchIcon } from '@patternfly/react-icons'
 import { ActionsColumn, IAction, SortByDirection, TableComposable, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table'
 import { ThSortType } from '@patternfly/react-table/dist/esm/components/Table/base'
 import useResizeObserver from '@react-hook/resize-observer'
-import { DateTime } from 'luxon'
 import {
-    CSSProperties,
     Dispatch,
     Fragment,
     MouseEvent,
@@ -59,18 +27,15 @@ import {
     useRef,
     useState,
 } from 'react'
-import { Link } from 'react-router-dom'
-import { BulkSelector } from './components/BulkSelector'
-import { IconWrapper } from './components/IconWrapper'
-import { getPatternflyColor, PatternFlyColor } from './components/patternfly-colors'
-import { SingleSelect2 } from './components/SingleSelect'
 import { useBreakpoint } from './components/useBreakPoint'
 import { PageBody } from './PageBody'
 import { useColumnModal } from './PageColumnModal'
 import { PageHeader, PageHeaderProps } from './PageHeader'
 import { PageLayout } from './PageLayout'
+import { PagePagination } from './PagePagination'
+import { IToolbarFilter, PageTableToolbar } from './PageToolbar'
 import { useSettings } from './Settings'
-import { IItemAction, isItemActionClick, ITypedAction, TypedActions, TypedActionType } from './TypedActions'
+import { IItemAction, isItemActionClick, ITypedAction, TypedActionType } from './TypedActions'
 
 export type TablePageProps<T extends object> = PageHeaderProps & PageTableProps<T> & { error?: Error }
 
@@ -598,433 +563,4 @@ export interface ITableColumn<T extends object> {
      * @deprecated The method should not be used
      */
     sortFn?: (l: T, r: T) => number
-}
-
-export type PagePaginationProps = {
-    itemCount?: number
-    page: number
-    perPage: number
-    setPage: (page: number) => void
-    setPerPage: (perPage: number) => void
-    style?: CSSProperties
-}
-
-export function PagePagination(props: PagePaginationProps) {
-    const { setPage, setPerPage } = props
-    const onSetPage = useCallback<OnSetPage>((_event, page) => setPage(page), [setPage])
-    const onPerPageSelect = useCallback<OnPerPageSelect>((_event, perPage) => setPerPage(perPage), [setPerPage])
-    const sm = useBreakpoint('md')
-    const settings = useSettings()
-    return (
-        <Pagination
-            variant={PaginationVariant.bottom}
-            itemCount={props.itemCount}
-            page={props.page}
-            perPage={props.perPage}
-            onSetPage={onSetPage}
-            onPerPageSelect={onPerPageSelect}
-            style={{
-                ...props.style,
-                borderTop: 'thin solid var(--pf-global--BorderColor--100)',
-                boxShadow: 'none',
-                zIndex: 301,
-                marginTop: -1,
-                paddingTop: sm ? 6 : undefined,
-                paddingBottom: sm ? 6 : undefined,
-                backgroundColor: settings.theme === 'dark' ? 'var(--pf-global--BackgroundColor--300)' : undefined,
-            }}
-            // titles={{ items: 'users' }}
-        />
-    )
-}
-
-export function toolbarActionsHaveBulkActions<T extends object>(actions?: ITypedAction<T>[]) {
-    if (!actions) return false
-    for (const action of actions) {
-        if (action.type === 'bulk') return true
-    }
-    return false
-}
-
-export interface IToolbarFilter {
-    key: string
-    label: string
-    type: string
-    query: string
-}
-
-export type IFilterState = Record<string, string[] | undefined>
-
-export type PagetableToolbarProps<T extends object> = { openColumnModal: () => void } & PageTableProps<T>
-
-export function PageTableToolbar<T extends object>(props: PagetableToolbarProps<T>) {
-    const {
-        itemCount,
-        page,
-        perPage,
-        setPage,
-        setPerPage,
-        toolbarFilters,
-        selectedItems,
-        filters,
-        setFilters,
-        clearAllFilters,
-        openColumnModal,
-    } = props
-
-    const sm = useBreakpoint('md')
-
-    let { toolbarActions } = props
-    toolbarActions = toolbarActions ?? []
-
-    const onSetPage = useCallback<OnSetPage>((_event, page) => setPage(page), [setPage])
-    const onPerPageSelect = useCallback<OnPerPageSelect>((_event, perPage) => setPerPage(perPage), [setPerPage])
-
-    const [filterValue, setFilterValue] = useState('')
-
-    const showSearchAndFilters = toolbarFilters !== undefined
-    const showToolbarActions = toolbarActions !== undefined && toolbarActions.length > 0
-
-    const showSelect =
-        selectedItems !== undefined && toolbarActions && toolbarActions.find((toolbarAction) => TypedActionType.bulk === toolbarAction.type)
-
-    const showToolbar = showSelect || showSearchAndFilters || showToolbarActions
-
-    const [selectedFilter, setSeletedFilter] = useState(() =>
-        toolbarFilters ? (toolbarFilters?.length > 0 ? toolbarFilters[0].key : '') : ''
-    )
-
-    const settings = useSettings()
-
-    if (!showToolbar) {
-        return <Fragment />
-    }
-
-    if (itemCount === undefined) {
-        return (
-            <Toolbar
-                style={{
-                    borderBottom: 'thin solid var(--pf-global--BorderColor--100)',
-                    paddingBottom: sm ? undefined : 8,
-                    paddingTop: sm ? undefined : 8,
-                    backgroundColor: settings.theme === 'dark' ? 'var(--pf-global--BackgroundColor--300)' : undefined,
-                }}
-            >
-                <ToolbarContent>
-                    <ToolbarItem style={{ width: '100%' }}>
-                        <Skeleton height="36px" />
-                    </ToolbarItem>
-                </ToolbarContent>
-            </Toolbar>
-        )
-    }
-
-    return (
-        <Toolbar
-            clearAllFilters={clearAllFilters}
-            style={{
-                borderBottom: 'thin solid var(--pf-global--BorderColor--100)',
-                paddingBottom: sm ? undefined : 8,
-                paddingTop: sm ? undefined : 8,
-                backgroundColor: settings.theme === 'dark' ? 'var(--pf-global--BackgroundColor--300)' : undefined,
-            }}
-        >
-            <ToolbarContent>
-                {showSelect && (
-                    <ToolbarGroup>
-                        <ToolbarItem variant="bulk-select">
-                            <BulkSelector {...props} />
-                        </ToolbarItem>
-                    </ToolbarGroup>
-                )}
-                {toolbarFilters && toolbarFilters.length > 0 && (
-                    <ToolbarToggleGroup toggleIcon={<FilterIcon />} breakpoint="md">
-                        <ToolbarGroup variant="filter-group">
-                            <ToolbarItem>
-                                <SingleSelect2 onChange={setSeletedFilter} value={selectedFilter}>
-                                    {toolbarFilters.map((filter) => (
-                                        <SelectOption key={filter.key} value={filter.key}>
-                                            <Flex
-                                                spaceItems={{ default: 'spaceItemsNone' }}
-                                                alignItems={{ default: 'alignItemsCenter' }}
-                                                flexWrap={{ default: 'nowrap' }}
-                                            >
-                                                <FlexItem style={{ paddingLeft: 4, paddingRight: 8 }}>
-                                                    <FilterIcon />
-                                                </FlexItem>
-                                                <FlexItem>{filter.label}</FlexItem>
-                                            </Flex>
-                                        </SelectOption>
-                                    ))}
-                                </SingleSelect2>
-                            </ToolbarItem>
-                            <ToolbarItem>
-                                <ToolbarTextFilter
-                                    value={filterValue}
-                                    setValue={setFilterValue}
-                                    addFilter={(value: string) => {
-                                        let values = filters?.[selectedFilter]
-                                        if (!values) values = []
-                                        if (!values.includes(value)) values.push(value)
-                                        setFilters?.({ ...filters, [selectedFilter]: values })
-                                    }}
-                                />
-                            </ToolbarItem>
-                            {toolbarFilters.map((filter) => {
-                                const values = filters?.[filter.key] ?? []
-                                return (
-                                    <ToolbarFilter
-                                        key={filter.label}
-                                        categoryName={filter.label}
-                                        chips={values}
-                                        deleteChip={(_group, value) => {
-                                            setFilters?.((filters) => {
-                                                const newState = { ...filters }
-                                                value = typeof value === 'string' ? value : value.key
-                                                let values = filters[filter.key]
-                                                if (values) {
-                                                    values = values.filter((v) => v !== value)
-                                                    if (values.length === 0) {
-                                                        delete newState[filter.key]
-                                                    } else {
-                                                        newState[filter.key] = values
-                                                    }
-                                                }
-                                                return newState
-                                            })
-                                        }}
-                                        deleteChipGroup={() => {
-                                            setFilters?.((filters) => {
-                                                const newState = { ...filters }
-                                                delete newState[filter.key]
-                                                return newState
-                                            })
-                                        }}
-                                        showToolbarItem={false}
-                                    >
-                                        <></>
-                                    </ToolbarFilter>
-                                )
-                            })}
-                        </ToolbarGroup>
-                    </ToolbarToggleGroup>
-                )}
-
-                {/* Action Buttons */}
-                <ToolbarGroup variant="button-group" style={{ zIndex: 302 }}>
-                    <TypedActions actions={toolbarActions} selectedItems={selectedItems} wrapper={ToolbarItem} />
-                    <ToolbarItem>
-                        <Tooltip content={'Manage columns'}>
-                            <Button variant="plain" icon={<ColumnsIcon />} onClick={openColumnModal} />
-                        </Tooltip>
-                    </ToolbarItem>
-                </ToolbarGroup>
-
-                {/* {toolbarButtonActions.length > 0 && <ToolbarGroup variant="button-group">{toolbarActionButtons}</ToolbarGroup>} */}
-                {/* <ToolbarGroup variant="button-group">{toolbarActionDropDownItems}</ToolbarGroup> */}
-
-                {/* Pagination */}
-                <ToolbarItem variant="pagination" visibility={{ default: 'hidden', xl: 'visible' }}>
-                    <Pagination
-                        variant={PaginationVariant.top}
-                        isCompact
-                        itemCount={itemCount}
-                        perPage={perPage}
-                        page={page}
-                        onSetPage={onSetPage}
-                        onPerPageSelect={onPerPageSelect}
-                        style={{ marginTop: -8, marginBottom: -8 }}
-                    />
-                </ToolbarItem>
-            </ToolbarContent>
-        </Toolbar>
-    )
-}
-
-function ToolbarTextFilter(props: { value: string; setValue: (value: string) => void; addFilter: (value: string) => void }) {
-    const { value, setValue } = props
-    // const ref = useRef<HTMLInputElement>()
-    return (
-        <InputGroup>
-            <TextInputGroup style={{ minWidth: 220 }}>
-                <TextInputGroupMain
-                    // ref={ref}
-                    value={value}
-                    onChange={setValue}
-                    onKeyUp={(event) => {
-                        if (value && event.key === 'Enter') {
-                            props.addFilter(value)
-                            setValue('')
-                            // ref.current?.focus() // Does not work because PF does not expose ref
-                        }
-                    }}
-                />
-                {value !== '' && (
-                    <TextInputGroupUtilities>
-                        <Button
-                            variant="plain"
-                            aria-label="add filter"
-                            onClick={() => setValue('')}
-                            style={{ opacity: value ? undefined : 0 }}
-                            // tabIndex={value ? undefined : -1}
-                            tabIndex={-1}
-                        >
-                            <TimesIcon />
-                        </Button>
-                    </TextInputGroupUtilities>
-                )}
-            </TextInputGroup>
-
-            {!value ? (
-                <Button variant={'control'} aria-label="add filter">
-                    <SearchIcon />
-                </Button>
-            ) : (
-                <Button
-                    variant={value ? 'primary' : 'control'}
-                    aria-label="add filter"
-                    onClick={() => {
-                        props.addFilter(value)
-                        setValue('')
-                    }}
-                >
-                    <ArrowRightIcon />
-                </Button>
-            )}
-        </InputGroup>
-    )
-}
-
-export function Labels(props: { labels: string[] }) {
-    return (
-        <LabelGroup numLabels={999} style={{ flexWrap: 'nowrap' }}>
-            {props.labels.map((label) => (
-                <Label key={label}>{label}</Label>
-            ))}
-        </LabelGroup>
-    )
-}
-
-export function DateCell(props: { value: number | string }) {
-    const date = new Date(props.value)
-    return (
-        <Split hasGutter>
-            <SplitItem>{date.toLocaleDateString()}</SplitItem>
-            <SplitItem>{date.toLocaleTimeString()}</SplitItem>
-        </Split>
-    )
-}
-
-export function TextCell(props: {
-    icon?: ReactNode
-    text?: string
-    iconSize?: 'sm' | 'md' | 'lg'
-    to?: string
-    onClick?: () => void
-    textColor?: PatternFlyColor
-}) {
-    return (
-        <Split>
-            {props.icon && (
-                <SplitItem>
-                    <IconWrapper size={props.iconSize ?? 'md'}>{props.icon}</IconWrapper>
-                </SplitItem>
-            )}
-            {props.to ? (
-                <SplitItem>
-                    {/* <Link to={props.to}>
-                        <Truncate position="middle" content={props.text ?? ''} />
-                    </Link> */}
-                    <Tooltip content={props.text ?? ''}>
-                        <div
-                            style={{
-                                maxWidth: 300,
-                                whiteSpace: 'nowrap',
-                                textOverflow: 'ellipsis',
-                                overflow: 'hidden',
-                            }}
-                        >
-                            <Link to={props.to}>{props.text ?? ''}</Link>
-                        </div>
-                    </Tooltip>
-                </SplitItem>
-            ) : props.onClick !== undefined ? (
-                <SplitItem onClick={props.onClick}>
-                    <Button
-                        variant="link"
-                        isInline
-                        style={{
-                            whiteSpace: 'nowrap',
-                        }}
-                    >
-                        {props.text}
-                    </Button>
-                </SplitItem>
-            ) : (
-                <SplitItem style={{ color: props.textColor ? getPatternflyColor(props.textColor) : undefined }}>
-                    <Tooltip content={props.text ?? ''}>
-                        <div
-                            style={{
-                                maxWidth: 300,
-                                whiteSpace: 'nowrap',
-                                textOverflow: 'ellipsis',
-                                overflow: 'hidden',
-                            }}
-                        >
-                            {props.text ?? ''}
-                        </div>
-                    </Tooltip>
-                </SplitItem>
-            )}
-        </Split>
-    )
-}
-
-export function CopyCell(props: { text?: string; minWidth?: number }) {
-    if (!props.text) return <></>
-    return (
-        <ClipboardCopy
-            hoverTip="Copy"
-            clickTip="Copied"
-            variant="inline-compact"
-            style={{ display: 'flex', flexWrap: 'nowrap', borderRadius: 4 }}
-            onCopy={() => {
-                void navigator.clipboard.writeText(props.text ?? '')
-            }}
-        >
-            <Truncate content={props.text} style={{ minWidth: props.minWidth }} />
-        </ClipboardCopy>
-    )
-}
-
-export function SinceCell(props: { value: string | undefined | null; author?: string; onClick?: () => void; t?: (t: string) => string }) {
-    let { t } = props
-    t = t ? t : (t: string) => t
-    const { author, onClick } = props
-    const [dateTime, setDateTime] = useState<string | null>(null)
-    useEffect(() => {
-        if (props.value) {
-            setDateTime(DateTime.fromISO(props.value).toRelative())
-        }
-        const timeout = setInterval(() => {
-            if (props.value) {
-                setDateTime(DateTime.fromISO(props.value).toRelative())
-            }
-        }, 1000)
-        return () => clearTimeout(timeout)
-    }, [props.value])
-    if (props.value === undefined) return <></>
-    return (
-        <span style={{ whiteSpace: 'nowrap' }}>
-            {dateTime}
-            {author && <span>&nbsp;{t('by')}&nbsp;</span>}
-            {onClick ? (
-                <Button variant="link" isInline onClick={onClick}>
-                    {author}
-                </Button>
-            ) : (
-                <span>{author}</span>
-            )}
-        </span>
-    )
 }
