+++ framework/index.ts
@@ -1,5 +1,4 @@
 export * from './BulkActionDialog'
-export * from './PageCatalog'
 export * from './components/BulkSelector'
 export * from './components/Collapse'
 export * from './components/Details'
@@ -7,6 +6,8 @@ export * from './components/Help'
 export * from './components/patternfly-colors'
 export * from './components/useBreakPoint'
 export * from './PageBody'
+export * from './PageCatalog'
+export * from './PageCells'
 export * from './PageDialog'
 export * from './PageForm'
 export * from './PageFramework'
@@ -14,6 +15,7 @@ export * from './PageHeader'
 export * from './PageLayout'
 export * from './PageTable'
 export * from './PageTabs'
+export * from './PageToolbar'
 export * from './TypedActions'
 export * from './useSelectDialog'
 export * from './useTableItems'
